fix: resolve circular dependency and OAuth2 OIDC ClassCastException
- Move PasswordEncoder bean from SecurityConfig to AppConfig to break
  the SecurityConfig → CustomOAuth2UserService → PasswordEncoder cycle
- Remove CustomOAuth2UserService from SecurityConfig constructor; inject
  it as a @Bean method parameter in securityFilterChain instead
- Handle DefaultOidcUser in OAuth2LoginSuccessHandler via instanceof
  pattern matching — Google uses OIDC flow so principal is not
  CustomOAuth2UserDetails but DefaultOidcUser; fall back to
  findOrCreateUser() when OIDC principal detected
- Expose CustomOAuth2UserService.findOrCreateUser() as public
- Update CLAUDE.md with OAuth2/OIDC flow notes and env var docs
