+++ CLAUDE.md
@@ -36,22 +36,35 @@ Spring Boot 3.5.7 REST API (Java 21) with JWT authentication, MongoDB persistenc
 **Layer structure:**
 - `controller/` — `AuthController` (public: `/auth/register`, `/auth/login`), `AdminController` (protected: `/user/profile`, `/admin/secret`, `/public/hello`)
 - `security/` — `JwtService` (HMAC-SHA256 token generation/validation), `JwtAuthenticationFilter`, `CustomUserDetailsService`, `CustomUserDetails`
-- `config/` — `SecurityConfig` (filter chain, CORS, CSRF-off, stateless sessions, `@PreAuthorize` method security), `AppConfig`
-- `entity/` — `User` and `Role` MongoDB documents; `Role` linked via `@DBRef`
-- `repository/` — Spring Data MongoDB repositories; `UserRepository.findByUsername()`
+- `config/` — `SecurityConfig` (filter chain, CORS, CSRF-off, stateless sessions, `@PreAuthorize` method security), `AppConfig` (`PasswordEncoder` bean + seed admin user)
+- `security/` — `CustomOAuth2UserService` (non-OIDC OAuth2 user load/create), `CustomOidcUserService`, `OAuth2LoginSuccessHandler`, `CookieOAuth2AuthorizationRequestRepository`
+- `entity/` — `User` and `Role` MongoDB documents; `Role` linked via `@DBRef`; `User` có `provider` + `providerId` cho OAuth2
+- `repository/` — Spring Data MongoDB repositories; `UserRepository.findByUsername()`, `findByEmail()`, `findByProviderAndProviderId()`
 - `dto/` — `RegisterRequest`, `LoginRequest`, `AuthResponse`, `UserDto`
 
 **Authorization model:** Role-based via Spring `@PreAuthorize`. Roles stored in MongoDB `Role` collection and referenced from `User`.
+
+**OAuth2 / Google login flow:**
+- Google dùng OIDC (có scope `openid`) → Spring Security gọi `OidcUserService` mặc định, principal là `DefaultOidcUser`
+- `OAuth2LoginSuccessHandler` dùng `instanceof` pattern để handle cả 2 loại principal: `CustomOAuth2UserDetails` (non-OIDC) và `DefaultOidcUser` (OIDC/Google)
+- Khi gặp `DefaultOidcUser`, gọi `CustomOAuth2UserService.findOrCreateUser()` để tìm/tạo user trong MongoDB
+- `PasswordEncoder` bean đặt trong `AppConfig` (không phải `SecurityConfig`) để tránh circular dependency: `SecurityConfig` → `CustomOAuth2UserService` → `PasswordEncoder` → `SecurityConfig`
+- Redirect sau login: `{frontend-redirect-uri}?token={jwt}&username={username}`
 
 ## Configuration
 
 `src/main/resources/application.yml` contains MongoDB URI and JWT settings. Override via environment variables in production:
 
 ```yaml
-spring.data.mongodb.uri    # MongoDB connection
-app.jwt.secret             # 32-char HMAC secret
-app.jwt.expiration-ms      # Token TTL (default: 86400000 = 24h)
+spring.data.mongodb.uri                          # MongoDB connection
+app.jwt.secret                                   # 32-char HMAC secret
+app.jwt.expiration-ms                            # Token TTL (default: 86400000 = 24h)
+GOOGLE_CLIENT_ID                                 # Google OAuth2 client ID
+GOOGLE_CLIENT_SECRET                             # Google OAuth2 client secret
+app.oauth2.frontend-redirect-uri                 # Frontend callback URL (e.g. http://localhost:5174/oauth2/callback)
 ```
+
+**Google Cloud Console:** Authorized redirect URI phải là `{baseUrl}/api/login/oauth2/code/google` (chú ý prefix `/api` theo context-path).
 
 ## Git Workflows
