+++ src/main/resources/application.yml
@@ -11,13 +11,13 @@ spring:
       client:
         registration:
           google:
-            client-id: ${GOOGLE_CLIENT_ID:your-google-client-id}
-            client-secret: ${GOOGLE_CLIENT_SECRET:your-google-client-secret}
+            client-id: ${GOOGLE_CLIENT_ID}
+            client-secret: ${GOOGLE_CLIENT_SECRET}
             scope:
               - openid
               - profile
               - email
-            redirect-uri: "{baseUrl}/login/oauth2/code/{registrationId}"
+            redirect-uri: ${GOOGLE_REDIRECT_URI:http://localhost:8080/api/login/oauth2/code/google}
         provider:
           google:
             user-name-attribute: sub
@@ -26,4 +26,4 @@ app:
     secret: ${JWT_SECRET:12345678901234567890123456789012}
     expiration-ms: ${JWT_EXPIRATION_MS:86400000}
   oauth2:
-    frontend-redirect-uri: http://localhost:3000/oauth2/callback
+    frontend-redirect-uri: ${FRONTEND_REDIRECT_URI:http://localhost:5174/oauth2/callback}
